+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/NabHibernateCommonConfig.java
@@ -24,7 +24,6 @@ import ru.hh.nab.hibernate.adapter.NabHibernatePersistenceProvider;
 import ru.hh.nab.hibernate.events.EventListenerRegistryPropagator;
 import ru.hh.nab.hibernate.service.NabServiceContributor;
 import ru.hh.nab.hibernate.service.ServiceSupplier;
-import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
 import ru.hh.nab.jpa.NabJpaCommonConfig;
 
 @Configuration
@@ -38,7 +37,6 @@ import ru.hh.nab.jpa.NabJpaCommonConfig;
     ExecuteOnDataSourceBeanPostProcessor.class,
     DataSourcesReadyTarget.class,
     TransactionalScope.class,
-    ExecuteOnDataSourceTransactionCallbackFactoryImpl.class,
 })
 public class NabHibernateCommonConfig {
