+++ nab-jpa/src/main/java/ru/hh/nab/jpa/NabJpaCommonConfig.java
@@ -9,15 +9,20 @@ import javax.sql.DataSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.context.annotation.Import;
 import org.springframework.orm.jpa.JpaVendorAdapter;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.SharedEntityManagerCreator;
 import org.springframework.orm.jpa.persistenceunit.PersistenceManagedTypes;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
+import ru.hh.nab.jpa.aspect.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
 
 @Configuration
 @EnableTransactionManagement(order = 0)
 @EnableAspectJAutoProxy
+@Import({
+    ExecuteOnDataSourceTransactionCallbackFactoryImpl.class,
+})
 public class NabJpaCommonConfig {
 
   @Bean
