+++ nab-jpa/src/main/java/ru/hh/nab/jpa/aspect/ExecuteOnDataSourceTransactionCallback.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.jpa.aspect;
 
 import jakarta.persistence.CacheRetrieveMode;
 import jakarta.persistence.CacheStoreMode;
